# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MangaGramm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/series:
    get:
      operationId: getUserSeries
      tags: [users]
      summary: Get series by user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
  /users/me/update:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /comments:
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/chapter/{chapterId}:
    get:
      operationId: getChapterComments
      tags: [comments]
      summary: Get comments for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /likes/toggle:
    post:
      operationId: toggleLike
      tags: [likes]
      summary: Toggle like on a series or chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleLikeBody"
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /likes/check:
    get:
      operationId: checkLike
      tags: [likes]
      summary: Check if user liked something
      parameters:
        - name: targetType
          in: query
          required: true
          schema:
            type: string
            enum: [series, chapter]
        - name: targetId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /follows/{authorId}:
    post:
      operationId: toggleFollow
      tags: [follows]
      summary: Toggle follow on an author
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
    get:
      operationId: checkFollow
      tags: [follows]
      summary: Check if following
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
...
  /follows/{userId}/followers:
    get:
      operationId: getUserFollowers
      tags: [follows]
      summary: Get user followers
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /favorites/toggle:
    post:
      operationId: toggleFavorite
      tags: [favorites]
      summary: Toggle favorite on a series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleFavoriteBody"
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
...
  /favorites/check/{seriesId}:
    get:
      operationId: checkFavorite
      tags: [favorites]
      summary: Check if series is favorited
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorite status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          type: string
          enum: [author, reader, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - role
        - createdAt
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
        role:
          type: string
      required:
        - id
        - username
        - role
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          type: string
        seriesCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
        totalViews:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - role
        - seriesCount
        - followersCount
        - followingCount
        - totalViews
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatar:
          type: string
    Series:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        bannerImage:
          type: string
          nullable: true
        type:
          type: string
          enum: [manga, webtoon, comic]
        status:
          type: string
          enum: [ongoing, completed, hiatus]
        genres:
          type: array
          items:
            type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        chapterCount:
          type: integer
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        mature:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - type
        - status
        - authorId
        - authorName
        - chapterCount
        - viewCount
        - likeCount
        - createdAt
        - updatedAt
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        userAvatar:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - chapterId
        - userId
        - username
        - content
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        chapterId:
          type: integer
        content:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - chapterId
        - content
    ToggleLikeBody:
      type: object
      properties:
        targetType:
          type: string
          enum: [series, chapter]
        targetId:
          type: integer
      required:
        - targetType
        - targetId
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required:
        - liked
        - likeCount
    FollowResponse:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
      required:
        - following
        - followersCount
    ToggleFavoriteBody:
      type: object
      properties:
        seriesId:
          type: integer
      required:
        - seriesId
    FavoriteResponse:
      type: object
      properties:
        favorited:
          type: boolean
        favoriteCount:
          type: integer
      required:
        - favorited
        - favoriteCount
...