# lib/api-zod/src/generated/types/createCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MangaGramm API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommentBody {
  chapterId: number;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
}