# lib/api-zod/src/generated/types/toggleLikeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MangaGramm API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ToggleLikeBodyTargetType } from "./toggleLikeBodyTargetType";

export interface ToggleLikeBody {
  targetType: ToggleLikeBodyTargetType;
  targetId: number;
}